+++ tests/testthat/test-with_test_id.R
@@ -1,0 +1,9 @@
+test_that("with_test_id works", {
+  tag <- shiny::div()
+  test_id <- "test-id"
+  tag_with_test_id <- with_test_id(tag, test_id)
+  expect_equal(
+    tag_with_test_id$attribs$`data-testid`,
+    test_id
+  )
+})