+++ NEWS.md
@@ -1,5 +1,7 @@
 # pw (development version)
 
+* `with_test_id()` adds a `data-testid` attribute to a shiny tag. This is useful for testing with `getByTestId` in Playwright.
+
 * `pw_show_report()` can be called to open the test report
 
 * `pw_codegen()` can be called to launch the code generator